refactor: relocate decrypt function for global usage
