+++ services/adminService.js
@@ -5,6 +5,7 @@ const AES_SECRET_KEY = process.env.AES_SECRET_KEY;
 const AES_IV = process.env.AES_IV;
 import { parse } from 'csv-parse';
 import fs, { unlink } from 'fs';
+import { decrypt } from '../middleware/decryption.js';
 
 const requiredColumns = ['role_name', 'department_name', 'user_name', 'password', 'workstation', 'email'];
 
@@ -17,22 +18,6 @@ const encrypt = (data) => {
 
 const hash = async (data) => {
   return await bcrypt.hash(data, 10);
-}
-
-const decrypt = (encryptedData) => {
-  try {
-    if (!encryptedData || typeof encryptedData !== 'string') {
-      return encryptedData; 
-    }
-
-    const decipher = crypto. createDecipheriv('aes-256-cbc', Buffer.from(AES_SECRET_KEY), Buffer.from(AES_IV));
-    let decrypted = decipher.update(encryptedData, 'base64', 'utf8');
-    decrypted += decipher.final('utf8');
-    return decrypted;
-  } catch (error) {
-    console.error('Decryption error:', error);
-    return encryptedData; 
-  }
 }
 
 /*
